Renamed migration package from Mysql to Mariadb
Resolves schema creation issue for mariadb
